Edges

Get edges from data.frame of tweets.

gt_edges_from_text(data, id, source, tweets, ...)

gt_edges_from_text_(data, id = "status_id", source = "screen_name",
  tweets = "text", ...)

gt_edges(data, source, target, ..., tl = TRUE)

gt_co_edges(data, col, tl = TRUE)

Arguments

data
Data.frame of tweets, usually returned by the rtweet package.

id
tweets unique id.

source
Author of tweets.

tweets
Column containing tweets.

...
any other column name, see examples.

target
Edges target.

tl
Set to TRUE to convert hashtags to lower case.

col
Column containing co-mentions.

Value

An object of class graphTweets.

Details

The tl arguments stands for tolower and allows converting the #hashtags to lower case as these often duplicated, i.e.: #python #Python.

Functions

gt_edges - Build networks of users.

gt_co_edges - Build networks of users to hashtags.

Examples

# simulate dataset
tweets <- data.frame(
  text = c("I tweet @you about @him and @her",
           "I tweet @me about @you"),
  screen_name = c("me", "him"),
  retweet_count = c(19, 5),
  status_id = c(1, 2),
  hashtags = c("rstats", "Python"),
  stringsAsFactors = FALSE
)

tweets %>%
  gt_edges_from_text(status_id, screen_name, text)

#> Warning: package 'bindrcpp' was built under R version 3.4.4

#> $edges
#> # A tibble: 5 x 2
#>   source target
#>   <chr>  <chr> 
#> 1 me     you   
#> 2 me     him   
#> 3 me     her   
#> 4 him    me    
#> 5 him    you   
#>